Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240511/합과 평균/sum-and-avg.cpp
@@ -10,7 +10,7 @@ int main() {
     float sum=a+b;
     float average= sum/2;
 
-    cout <<sum<<" "<< average;
+    cout <<sum<<" "<< (float)average;
 
     return 0;
 }
